Issue 536: Refactored urls in tests
